+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * New interactive command `clojure-mode-report-bug`.
 * New interactive command `clojure-view-guide`.
 * New interactive command `clojure-view-reference-section`.
+* New interactive command `clojure-view-cheatsheet`.
 * Make the refactoring keymap prefix customizable via `clojure-refactor-map-prefix`.
 
 ## 5.5.2 (2016-08-03)